Zoyas changes

+++ include/Inertial.hpp
@@ -5,4 +5,4 @@
 #include <cstdint>
 
 
-std::shared_ptr<okapi::ChassisController> build_PID (okapi::MotorGroup left_motor, okapi::MotorGroup right_motor, int inertial1, int inertial2);
+std::shared_ptr<okapi::ChassisController> build_PID (okapi::MotorGroup left_motor, okapi::MotorGroup right_motor, int8_t inertial1, int8_t inertial2);
